pyproject.toml
[project]
name = "treasury"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.3",
]
